pyproject.toml
[project]
name = "convert-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow-heif>=0.13.0",
    "pillow>=10.0.0",
]
